requirements.txt
psutil==5.9.6
joblib==1.3.2
pandas==2.1.1
scikit-learn==1.3.2
python-telegram-bot==20.7
scapy==2.5.0
